Fixed issues in Services

+++ Resonance/Services/Interfaces/ICourseService.cs
@@ -5,9 +5,9 @@ namespace ResoClassAPI.Services.Interfaces
     public interface ICourseService
     {
         Task<List<CourseDto>> GetAllCourses();
-        Task<CourseDto> GetCourse(int courseId);
+        Task<CourseDto> GetCourse(long courseId);
         Task<long> CreateCourse(CourseDto newItem);
         Task<bool> UpdateCourse(CourseDto updatedItem);
-        Task<bool> DeleteCourse(int courseId);
+        Task<bool> DeleteCourse(long courseId);
     }
 }
