+++ Resonance/Services/Interfaces/ITopicService.cs
@@ -5,9 +5,9 @@ namespace ResoClassAPI.Services.Interfaces
     public interface ITopicService
     {
         Task<List<TopicDto>> GetAllTopics();
-        Task<TopicDto> GetTopic(int topicId);
+        Task<TopicDto> GetTopic(long topicId);
         Task<long> CreateTopic(TopicDto newItem);
         Task<bool> UpdateTopic(TopicDto updatedItem);
-        Task<bool> DeleteTopic(int topicId);
+        Task<bool> DeleteTopic(long topicId);
     }
 }
