+++ Resonance/Services/Interfaces/IVideoService.cs
@@ -5,9 +5,9 @@ namespace ResoClassAPI.Services.Interfaces
     public interface IVideoService
     {
         Task<List<VideoDto>> GetAllVideos();
-        Task<VideoDto> GetVideo(int topicId);
+        Task<VideoDto> GetVideo(long topicId);
         Task<long> CreateVideo(VideoDto newItem);
         Task<bool> UpdateVideo(VideoDto updatedItem);
-        Task<bool> DeleteVideo(int topicId);
+        Task<bool> DeleteVideo(long topicId);
     }
 }
